Update bindgen to version 0.55
This version of bindgen had a change in the way that it outputs
union fields with names of builtin types (a field named u64 is now
named u64_) so this commit also updates the c2rust build scripts
to take that into account.

As I was fixing up the build scripts I also updated them to fix
some issues that I hit while trying to get them to work.