Add linux CI via github actions (#48)
Travis CI no longer seems to be running. This commit sets up github actions
to build/test xed-sys on linux. It mainly mirrors the tests run by appveyor.

The main change that this commit makes otherwise is to not run clippy
during CI. There's too many false positive warnings in CI due to all the generated
code so it turns out to not be useful here.