Fix cross-compilation
When cross-compiling, the `--toolchain` argument needs to be passed to
XED's `mfile.py`. This ensures that XED builds its object files with the
correct toolchain binaries. When compiling with `cross build`, the
CROSS_TOOLCHAIN_PREFIX environment variable is automatically provided.
This patch ensures that the environment variable is passed to `mfile.py`
through the `--toolchain` argument.